Multithreaded World Ticking
This push finally brings Multithreaded WorldTicking to the public. This code makes it so that worlds tick in parallel with eachother on different threads.

In VanillaMC, it uses the "Ping-Pong" method for ticking worlds, aka ticking 1 world at a time. This can cause a ton of performance issues for servers with multiple worlds, or busy ones. Canvas now makes it so that it ticks in "batches" on different threads to tick them all at once to save on tick times.

Performance has improved DRASTICALLY with this new method, and best of all, it doesnt break plugin compatibility!!

Some other notable changes are:
- patches have been merged and formatted to be better
- Ignite has been updated in upstream, reducing a lot of file size and improving the launcher
- MixinPlugins now exist to try and create a safer way of making plugins that utilize mixin(docs on the example plugin)