Fix spamming entities through portal creating tons of portals
When a portal is just constructed, you can spam entities through the portal. If you do this, every single entity will try and create a new portal. As a result, TONS of portals get created.

This is caused by when the 1st portal exit is loaded, the poi record isn't placed, so every single one following that fail the 2nd "is the portal existing now" check, so every single entity going through that portal tries to load a new portal. By placing them manually, we then ensure the 1st portal creation adds a poi record, updates it, then every following entity passes the 2nd "is the portal existing now" check