Reload player on teleport cross-region and portaling
In Folia's teleport rewrite, there are 3 kinds of teleports. Sync-Region, Cross-Region, and Cross-Dimensional. Essentially, in cross-region and cross-dimensional(portaling), it removes the player from the world. This causes a visual effect where the client looks like it is in the void... which isn't very good.

To try and clean this up, this commit introduces sending packets to the client that display the "Loading terrain..." screen, which also matches the destination dimension. This calls on cross-region and portaling teleport types, which make it a more seemless teleport, rather than just dropping into the void... And also makes it more align with Vanilla