Reduce isTickThreadFor lookups during playerlist save
With high player counts, each region runs through every single player on the server and calls a tick thread check on them. This is kinda redundant, as we can literally just fetch the local players and call it a day. This makes it so if the interval isn't -1, which is only possible during region autosave, it will try and fetch the local players which is much more efficient than iterating over every single player on the server for autosaving the 1 single regions players.