Ensure the hooked entity is within same region as the fishing hook
Folia only checks if the owning player is in the same region as the hook. The hook can technically be hooked onto an entity in another region by some race conditions like if the entity teleports away, the region is split, etc.

This replaces the level check with a region check, which fixes any race condition from the hooked entity being thrown into a different region.