Use better scheduling and use safer iteration for waypoints rewrite
The waypoints rewrite had an issue where it could throw an AIOOB exception, which was originally thought to be a scheduling issue and/or a CME. It was later realized that the iteration during RegionThreadingWaypointManager#updatePlayer was incorrect, as during the iteration it could mutate the map directly, which can cause an AIOOB exception to be thrown.

This commit fixes this by changing RegionThreadingWaypointManager#updatePlayer to use a RegionThreadingWaypointManager.CallbackHolder system where the method would return a callback with a nullable Runnable that defined any mutations to the map during the iteration so it can be delayed post-iteration. Some scheduling changes were also made to be simplified by using ServerPlayer#scheduleToOrRun