Change default from 0.3ms to 0.1ms
It is unlikely that the server will be missing deadlines and such because of mid tick tasks. However, the possibility is still there. So I am making the buffer smaller, in the hopes of being able to process more mid-tick-tasks, while not missing deadlines

This also makes it so that the scheduler won't try and run mid-tick-tasks if the region is within the buffer range(default now 0.1ms).