Fix Folia crash with TamableAnimal
The tamable entity can, when trying to teleport to its owner, try and load chunks that may or may not be inside the region boundaries. If it does try and load chunks *outside* the region, the server crashes. So we do a quick check first now to ensure the chunk *is* loaded first, since we can get blocks outside the region boundaries if the chunk is loaded. If the chunk is loaded, we will teleport like normal. Otherwise, just don't even bother trying.