Remove sending respawn packet on async teleport
Generally, this feature was useful and a good QOL feature for 1.21.8 servers at the time. Now, this packet does something different, and no longer displays the screen that it was intended to display. As such, it has been decided to remove this feature to match with Folia, as I see no point in continuing to contain and maintain this in Canvas.